Working examples in tests client-web
