+++ client-web/package.json
@@ -30,15 +30,14 @@
   },
   "dependencies": {
     "@alkemio/client-lib": "^0.32.0",
-    "@alkemio/tests-lib": "^0.1.4",
     "@ory/kratos-client": "^1.2.1",
     "axios": "^1.7.9",
     "cross-env": "^7.0.3",
     "graphql": "^16.6.0",
     "graphql-type-json": "^0.3.2",
     "graphql-upload": "^17.0.0",
-    "supertest": "^7.0.0",
-    "graphql-ws": "^5.8.2"
+    "graphql-ws": "^5.8.2",
+    "supertest": "^7.0.0"
   },
   "devDependencies": {
     "@graphql-codegen/add": "^5.0.3",
